implemented motor, servo, logic engine debug functionality

+++ astromech/packet_arduino/motor_servo_control.h
@@ -9,7 +9,7 @@ void motor_setup();
 
 // These Functions Set the Speed of the Specified Motor/Servo
 // The Speed is Clamped to Values Between -100 and 100
-void control_motors_joystick(uint8_t ver_val, uint16_t hor_val);
+void control_motors_joystick(uint16_t ver_val, uint16_t hor_val);
 void change_motor_speed(uint8_t motor_num, int16_t speed);
 void set_servo_angle(uint8_t servo_num, int16_t speed);
